# docs/swagger.yaml
basePath: /
definitions:
...
  models.ErrorResponse:
    properties:
      error: {}
    type: object
...
  models.InfoFaq:
    properties:
      pregunta:
        type: string
      respuesta:
        type: string
    type: object
  models.InfoLocation:
    properties:
      ciudad:
        type: string
      coordenadas:
        type: string
      direccion:
        type: string
      horario:
        additionalProperties:
          type: string
        type: object
      nombre:
        type: string
    type: object
...
  models.SuccessResponse:
    properties:
      notice:
        type: string
      record: {}
    type: object
...
host: 127.0.0.1:7003
info:
  contact:
    email: juan9acosta@gmail.com
    name: Juan Acosta
    url: https://www.linkedin.com/in/juan-m-acosta-f-54219758/
  description: service in Go using Gin framework
  title: MiCuenta Service API
  version: "1.0"
paths:
...
  /info/accesibilidad:
    get:
      consumes:
      - application/json
      description: Retrieve Texto con accesibilidad
      parameters:
      - description: Tipo de plataforma
        enum:
        - movil
        - web
        in: query
        name: tipo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid Request or Incorrect Data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Texto con accesibilidad para la app movil
      tags:
      - Info
  /info/faqs:
    get:
      consumes:
      - application/json
      description: Retrieve listado de preguntas frecuentes
      parameters:
      - description: Tipo de plataforma
        enum:
        - movil
        - web
        in: query
        name: tipo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                record:
                  items:
                    $ref: '#/definitions/models.InfoFaq'
                  type: array
              type: object
        "400":
          description: Invalid Request or Incorrect Data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: listado de preguntas frecuentes
      tags:
      - Info
  /info/legal/privacy_policy:
    get:
      consumes:
      - application/json
      description: Retrieve Texto con politicas de privacidad
      parameters:
      - description: Tipo de plataforma
        enum:
        - movil
        - web
        in: query
        name: tipo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid Request or Incorrect Data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Texto con politicas de privacidad
      tags:
      - Info
  /info/legal/terminos_y_condiciones:
    get:
      consumes:
      - application/json
      description: Retrieve Texto con los terminos y condiciones
      parameters:
      - description: Tipo de plataforma
        enum:
        - movil
        - web
        in: query
        name: tipo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid Request or Incorrect Data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Texto con los terminos y condiciones
      tags:
      - Info
  /info/oficinas:
    get:
      consumes:
      - application/json
      description: Retrieve a list of oficinas de la empresa con su ubicacion
      parameters:
      - description: Tipo de plataforma
        enum:
        - movil
        - web
        in: query
        name: tipo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                record:
                  items:
                    $ref: '#/definitions/models.InfoLocation'
                  type: array
              type: object
        "400":
          description: Invalid Request or Incorrect Data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Listado de oficinas
      tags:
      - Info
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"